+++ test/ECJsonMD.test.ts
@@ -156,17 +156,17 @@ describe("ECJsonToMD", () => {
 
     it("should write the classes as a table", () => {
       // Check that the classes print into a table with the correct name, description, and type
-      assert.equal(lines[14], "|PropertyOne||int|");
+      assert.equal(lines[14], "|PropertyOne||int||");
 
       // Check that the classes print into a table with the correct name, description, and type
-      assert.equal(lines[27], "|PropertyOne|This is the first property of ClassOne|string|");
-      assert.equal(lines[28], "|PropertyTwo|This is the second property of ClassOne.|string|");
-      assert.equal(lines[29], "|PropertyThree|This is the third property of ClassOne|int|");
+      assert.equal(lines[27], "|PropertyOne|This is the first property of ClassOne|string|Json|");
+      assert.equal(lines[28], "|PropertyTwo|This is the second property of ClassOne.|string||");
+      assert.equal(lines[29], "|PropertyThree|This is the third property of ClassOne|int||");
 
       // Check that the classes print into a table with the correct name, description, and type
-      assert.equal(lines[56], "|PropertyOne|This is the first property of ClassTwo|int|");
-      assert.equal(lines[57], "|PropertyTwo|This is the second property of ClassTwo.|string|");
-      assert.equal(lines[58], "|PropertyThree|This is the third property of ClassTwo|int|");
+      assert.equal(lines[56], "|PropertyOne|This is the first property of ClassTwo|int||");
+      assert.equal(lines[57], "|PropertyTwo|This is the second property of ClassTwo.|string||");
+      assert.equal(lines[58], "|PropertyThree|This is the third property of ClassTwo|int||");
       });
 
     it("should write the name of the classes as h2", () => {
@@ -191,16 +191,16 @@ describe("ECJsonToMD", () => {
     });
 
     it("should write a property without a description", () => {
-      assert.equal(lines[14], "|PropertyOne||int|");
+      assert.equal(lines[14], "|PropertyOne||int||");
     });
 
     it("should put an empty table at the end of each property table", () => {
       assert.equal(lines[15],
-        "|                                      |                                   |            |");
+        "|            |                   |            |                        |");
       assert.equal(lines[30],
-        "|                                      |                                   |            |");
+        "|            |                   |            |                        |");
       assert.equal(lines[59],
-        "|                                      |                                   |            |");
+        "|            |                   |            |                        |");
 
       // Empty row at the end of a relationship table
       assert.equal(lines[50],
@@ -214,6 +214,11 @@ describe("ECJsonToMD", () => {
 
     it("should write the baseclass properly", () => {
       assert.equal(lines[42], "**Base class:** SchemaA:ClassOne");
+    });
+
+    it("should list the extended typename for properties", () => {
+      assert.equal(lines[27], "|PropertyOne|This is the first property of ClassOne|string|Json|");
+      assert.equal(lines[28], "|PropertyTwo|This is the second property of ClassOne.|string||");
     });
  });